+++ contracts/test/TestExecAccount.sol
@@ -50,7 +50,7 @@ contract TestExecAccountFactory {
 
     function createAccount(address owner, uint256 salt) public returns (address ret) {
         address addr = getAddress(owner, salt);
-        uint codeSize = addr.code.length;
+        uint256 codeSize = addr.code.length;
         if (codeSize > 0) {
             return addr;
         }
